parser-service-1/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-pubsub
lxml
requests
python-dotenv